[lib/host] Move things around to make them be in the right places
- Move validate_and_hash function to Memory class. The function itself
  is not refactored
- Move out unrelated variables from keystone_user.h
- Hash_util should go somewhere else, but for now, we treat it as a c++
  file
