+++ lib/host/include/hash_util.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+extern "C" {
+#include "./sha3.h"
+}
+
+typedef sha3_ctx_t hash_ctx_t;
+
+void hash_init(hash_ctx_t* hash_ctx);
+void hash_extend(hash_ctx_t* hash_ctx, const void* ptr, size_t len);
+void hash_extend_page(hash_ctx_t* hash_ctx, const void* ptr);
+void hash_finalize(void* md, hash_ctx_t* hash_ctx);
+void printHash(char *hash);
