+++ lib/host/include/keystone_user.h
@@ -25,14 +25,6 @@
 #define KEYSTONE_IOC_UTM_INIT \
   _IOR(KEYSTONE_IOC_MAGIC, 0x07, struct keystone_ioctl_create_enclave)
 
-#define RT_NOEXEC 0
-#define USER_NOEXEC 1
-#define RT_FULL 2
-#define USER_FULL 3
-#define UTM_FULL 4
-
-#define MDSIZE 64
-
 struct runtime_params_t {
   uintptr_t runtime_entry;
   uintptr_t user_entry;
