+++ lib/host/include/sha3.h
@@ -1,7 +1,5 @@
 // 19-Nov-11  Markku-Juhani O. Saarinen <mjos@iki.fi>
-
-#ifndef SHA3_H
-#define SHA3_H
+#pragma once
 
 #include <stdint.h>
 #include <stddef.h>
@@ -13,6 +11,8 @@
 #ifndef ROTL64
 #define ROTL64(x, y) (((x) << (y)) | ((x) >> (64 - (y))))
 #endif
+
+#define MDSIZE 64
 
 // state context
 typedef struct {
@@ -33,5 +33,3 @@ int sha3_final(void *md, sha3_ctx_t *c);    // digest goes to md
 
 // compute a sha3 hash (md) of given byte length from "in"
 void *sha3(const void *in, size_t inlen, void *md, int mdlen);
-
-#endif
